make rails_admin product name (i.e., "PDMP Spike") link to "/" instead of "/admin"
We need to do this so users can get out of rails_admin back to the
main application.  Had to copy the rails_admin application layout from
the gem to the app which might end up being a hassle in the long term
but there doesn't seem to be a way to configure that link.