silence Rack deprecation messages during tests
When we upgraded to Rails 3.2.11 it pulled in a new rack which printed
an annoying deprecation message during unit tests.  Removed the static
asset config for tests to quiet it.  In practice it won't make a
difference in test performance and it will keep me from going nuts
looking at spurious messages.