rearrange srpp section/document relationship
We used to have three sections with one document each, but that's
overly complex so now we've got one "srpp" section with three
documents in it.  Actually, I haven't removed the old sections or
routes yet, since I want to give the client implementations a chance
to migrate to the new layout.